+++ src/components/InitialMessage.tsx
@@ -1,0 +1,66 @@
+// SPDX-License-Identifier: Apache-2.0
+//
+// SPDX-FileCopyrightText: © 2025 Tenstorrent AI ULC
+
+import { Button, Dialog, DialogBody, DialogFooter, Intent } from '@blueprintjs/core';
+import { useState } from 'react';
+import getServerConfig from '../functions/getServerConfig';
+
+const InitialMessage = () => {
+    const [isDialogOpen, setIsDialogOpen] = useState(!shouldHideStartup());
+
+    const handleCloseDialog = () => {
+        createHideStartupCookie();
+        setIsDialogOpen(false);
+    };
+
+    return (
+        <Dialog
+            title='Welcome to TT-NN Visualizer'
+            icon='info-sign'
+            isOpen={isDialogOpen}
+            usePortal={false}
+            onClose={handleCloseDialog}
+        >
+            <DialogBody>
+                <p>
+                    Choose from our demo reports or upload your own generated with{' '}
+                    <a href='https://github.com/tenstorrent/tt-metal/'>TT-Metal</a> to visualize the data.
+                </p>
+
+                <p>
+                    Visualisation of <a href='https://github.com/tenstorrent/tt-npe'>TT-NPE</a> data is also supported
+                    and can be found in the NPE section of the app.
+                </p>
+            </DialogBody>
+
+            <DialogFooter
+                actions={
+                    <Button
+                        intent={Intent.PRIMARY}
+                        text='Close'
+                        onClick={handleCloseDialog}
+                    />
+                }
+            />
+        </Dialog>
+    );
+};
+
+const shouldHideStartup = (): boolean => {
+    const match = document.cookie.match(/(?:^|;\s*)hide-startup-information=([^;]*)/);
+
+    if (!getServerConfig()?.SERVER_MODE) {
+        return false;
+    }
+
+    return match ? match[1] === 'true' : false;
+};
+
+const createHideStartupCookie = () => {
+    const expires = new Date();
+    expires.setDate(expires.getDate() + 30);
+    document.cookie = `hide-startup-information=true; expires=${expires.toUTCString()}; path=/`;
+};
+
+export default InitialMessage;
