+++ src/routes/Home.tsx
@@ -2,27 +2,20 @@
 //
 // SPDX-FileCopyrightText: © 2025 Tenstorrent AI ULC
 
-import { useState } from 'react';
-import { Button, Callout, Dialog, DialogBody, DialogFooter, Icon, Intent } from '@blueprintjs/core';
+import { Callout, Icon, Intent } from '@blueprintjs/core';
 import { IconNames } from '@blueprintjs/icons';
 import LocalFolderSelector from '../components/report-selection/LocalFolderSelector';
 import RemoteSyncConfigurator from '../components/report-selection/RemoteSyncConfigurator';
 import 'styles/routes/Home.scss';
 import useClearSelectedBuffer from '../functions/clearSelectedBuffer';
 import getServerConfig from '../functions/getServerConfig';
+import InitialMessage from '../components/InitialMessage';
 
 function Home() {
     useClearSelectedBuffer();
 
-    const [isDialogOpen, setIsDialogOpen] = useState(!shouldHideStartup());
-
     const disableRemoteSync = !!getServerConfig()?.SERVER_MODE;
     const isDialogEnabled = !!getServerConfig()?.SERVER_MODE;
-
-    const handleCloseDialog = () => {
-        createHideStartupCookie();
-        setIsDialogOpen(false);
-    };
 
     return (
         <div className='home'>
@@ -80,55 +73,9 @@ function Home() {
                 </fieldset>
             </div>
 
-            {isDialogEnabled && (
-                <Dialog
-                    title='Welcome to TT-NN Visualizer'
-                    icon='info-sign'
-                    isOpen={isDialogOpen}
-                    usePortal={false}
-                    onClose={handleCloseDialog}
-                >
-                    <DialogBody>
-                        <p>
-                            Choose from our demo reports or upload your own generated with{' '}
-                            <a href='https://github.com/tenstorrent/tt-metal/'>TT-Metal</a> to visualize the data.
-                        </p>
-
-                        <p>
-                            Visualisation of <a href='https://github.com/tenstorrent/tt-npe'>TT-NPE</a> data is also
-                            supported and can be found in the NPE section of the app.
-                        </p>
-                    </DialogBody>
-
-                    <DialogFooter
-                        actions={
-                            <Button
-                                intent={Intent.PRIMARY}
-                                text='Close'
-                                onClick={handleCloseDialog}
-                            />
-                        }
-                    />
-                </Dialog>
-            )}
+            {isDialogEnabled && <InitialMessage />}
         </div>
     );
 }
-
-const shouldHideStartup = (): boolean => {
-    const match = document.cookie.match(/(?:^|;\s*)hide-startup-information=([^;]*)/);
-
-    if (!getServerConfig()?.SERVER_MODE) {
-        return false;
-    }
-
-    return match ? match[1] === 'true' : false;
-};
-
-const createHideStartupCookie = () => {
-    const expires = new Date();
-    expires.setDate(expires.getDate() + 30);
-    document.cookie = `hide-startup-information=true; expires=${expires.toUTCString()}; path=/`;
-};
 
 export default Home;
